requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
idna==3.3
python-nmap==0.7.1
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.12
